requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
efficientnet-pytorch
gdown
geos
geotiff
markdown
matplotlib
numpy
opencv-python-headless
pandas
Pillow
pip
pretrainedmodels
pytorch-lightning
rasterio
scipy
segmentation-models-pytorch
sklearn
streamlit
timm
torch
torchmetrics
torchvision
tqdm
xarray
xarray-spatial